requirements.txt
beautifulsoup4==4.12.3
cdx_toolkit==0.9.35
Cython~=3.0.0
dnspython~=1.15.0
exrex~=0.11.0
fake_useragent==1.5.1
fire~=0.5.0
googletrans~=3.0.0
idna==2.10
loguru~=0.7.0
lxml==4.9.3
prettytable==3.10.0
PyQt5~=5.15.7
PyQt5_sip==12.11.0
Requests==2.31.0
SQLAlchemy~=2.0.28
tenacity==8.2.3
tqdm~=4.66.2
treelib~=1.7.0
urllib3~=1.26.12
python-whois
tldextract~=5.1.1
bs4~=0.0.2
requests[socks]